LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/intrinsic-size/col-wrap-005-expected.txt


FAIL body > div 1 assert_equals:
<div style="display: flex; flex-flow: column wrap; width: max-content; height: 100px; background: green; position: relative;" data-expected-width="250">
  <div class="item" style="width: 100px;"></div>
  <div class="item" style="width: fit-content;" data-expected-width="150" data-offset-x="100">
    <!-- This item has min-content=75 and max-content=150. -->
    <div class="grandchild"></div>
    <div class="grandchild"></div>
  </div>
</div>
width expected 250 but got 150
